+++ jxls/src/main/java/org/jxls/common/ObjectPropertyAccess.java
@@ -4,7 +4,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Map;
 
-import org.apache.commons.beanutils.PropertyUtils;
+import org.apache.commons.beanutils2.PropertyUtils;
 import org.jxls.logging.JxlsLogger;
 
 public class ObjectPropertyAccess {